+++ scripts/LiveRouteOCR/$Recycle.Bin/S-1-5-21-97200220-2965608657-3582652261-1001/$RBM0GY7.js
@@ -1,6 +1,0 @@
-// electron/preload.cjs
-const { contextBridge } = require("electron");
-
-contextBridge.exposeInMainWorld("pokemmo", {
-  version: "1.2.5"
-});